use rainbow kit for wallet connection
