+++ example/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@ethersproject/abstract-provider": "^5.7.0",
     "@nft/chat": "file:..",
+    "@rainbow-me/rainbowkit": "^0.5.1",
     "next": "12.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0"
