--- example/types.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-  ethereum?: import('@metamask/providers').BaseProvider
-}
